Expanded app layer message object size field to support >255 bytes
Required for large message handling

+++ electricui.h
@@ -16,7 +16,7 @@ typedef void (*CallBackType)();            //callback with no data
 typedef struct {
     const char*   msgID;
     uint8_t       type;
-    uint8_t       size;
+    uint16_t      size;
     void          *payload;
 } euiMessage_t;
